OpenFinProvider now receives desktop context
