+++ src/client/src/ui/shell/OpenFinProvider.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { EnvironmentConsumer } from '../../main'
 import { OpenFin } from '../../services'
 import OpenFinChrome from './OpenFinChrome'
 
@@ -6,18 +7,23 @@ interface OpenFinShellProps {
   openFin: OpenFin
 }
 
-export default class OpenFinMainWindow extends React.Component<OpenFinShellProps> {
-  render() {
-    const { openFin } = this.props
-    return (
-      <OpenFinChrome
-        openFin={openFin}
-        minimize={openFin.minimize.bind(openFin)}
-        maximize={openFin.maximize.bind(openFin)}
-        close={openFin.close.bind(openFin)}
-      >
-        {this.props.children}
-      </OpenFinChrome>
-    )
-  }
-}
+const OpenFinMainWindow: React.SFC<OpenFinShellProps> = ({ openFin, children }) => (
+  <EnvironmentConsumer>
+    {isRunningOnDesktop =>
+      isRunningOnDesktop ? (
+        <OpenFinChrome
+          openFin={openFin}
+          minimize={openFin.minimize.bind(openFin)}
+          maximize={openFin.maximize.bind(openFin)}
+          close={openFin.close.bind(openFin)}
+        >
+          {children}
+        </OpenFinChrome>
+      ) : (
+        children
+      )
+    }
+  </EnvironmentConsumer>
+)
+
+export default OpenFinMainWindow
